+++ serial_interface/examples/c/Makefile
@@ -5,17 +5,17 @@ DEBUG   = 1
 C_LOC = $(realpath .)
 
 CHIMESFF_LOC=$(C_LOC)/../../../chimesFF/src
-CHIMESFF_SRC=$(CHIMESFF_LOC)/chimescalc.cpp
-CHIMESFF_HDR=$(CHIMESFF_LOC)/chimescalc.h
+CHIMESFF_SRC=$(CHIMESFF_LOC)/chimesFF.cpp
+CHIMESFF_HDR=$(CHIMESFF_LOC)/chimesFF.h
 
-chimescalc.o : $(CHIMESFF_SRC)
+chimesFF.o : $(CHIMESFF_SRC)
 	$(CXX) -c $(CHIMESFF_SRC) -I $(CHIMESFF_LOC)
 
 SERIAL_LOC=$(C_LOC)/../../src
-SERIAL_SRC=$(SERIAL_LOC)/chimescalc_serial.cpp
-SERIAL_HDR=$(SERIAL_LOC)/chimescalc_serial.h
+SERIAL_SRC=$(SERIAL_LOC)/serial_chimes_interface.cpp
+SERIAL_HDR=$(SERIAL_LOC)/serial_chimes_interface.h
 
-chimescalc_serial.o : $(SERIAL_SRC)
+serial_chimes_interface.o : $(SERIAL_SRC)
 	$(CXX) -c $(SERIAL_SRC) -I $(SERIAL_LOC) -I $(CHIMESFF_LOC)
 
 WRAPPER_LOC=$(C_LOC)/../../api
@@ -31,7 +31,7 @@ TEST_SRC=$(TEST_LOC)/main.c
 main.o: $(TEST_SRC)
 	$(CCOMP) -c $(TEST_SRC) -DDEBUG=${DEBUG}  -I $(WRAPPER_LOC) -I $(SERIAL_LOC) -I $(CHIMESFF_LOC)
 
-TEST_LNK = chimescalc.o chimescalc_serial.o chimescalc_serial_C.o main.o
+TEST_LNK = chimesFF.o serial_chimes_interface.o chimescalc_serial_C.o main.o
 
 test-C: $(TEST_LNK)
 	$(CXX) $(TEST_LNK) -o C_wrapper-serial_interface
@@ -44,8 +44,8 @@ clean-all:
 	rm -f C_wrapper-serial_interface
 
 all:
-	make chimescalc.o
-	make chimescalc_serial.o
+	make chimesFF.o
+	make serial_chimes_interface.o
 	make chimescalc_serial_C.o
 	make main.o
 	make test-C
