+++ serial_interface/examples/cpp/Makefile
@@ -3,8 +3,8 @@ CXX   = g++ -O3 -std=c++11
 DEBUG = 1
 
 CXX_LOC = $(realpath .)
-CXX_SRC = main.cpp $(CXX_LOC)/../../src/chimescalc_serial.cpp $(CXX_LOC)/../../../chimesFF/src/chimescalc.cpp
-CXX_HDR = $(CXX_LOC)/../../src/chimescalc_serial.h $(CXX_LOC)/../../../chimesFF/src/chimescalc.h
+CXX_SRC = main.cpp $(CXX_LOC)/../../src/serial_chimes_interface.cpp $(CXX_LOC)/../../../chimesFF/src/chimesFF.cpp
+CXX_HDR = $(CXX_LOC)/../../src/serial_chimes_interface.h $(CXX_LOC)/../../../chimesFF/src/chimesFF.h
 
 clean:
