+++ serial_interface/examples/python/Makefile
@@ -3,17 +3,17 @@ C_LOC = $(realpath .)
 CXX     = g++ -O3 -std=c++11  # -fPIC
 
 CHIMESFF_LOC=$(C_LOC)/../../../chimesFF/src
-CHIMESFF_SRC=$(CHIMESFF_LOC)/chimescalc.cpp
-CHIMESFF_HDR=$(CHIMESFF_LOC)/chimescalc.h
+CHIMESFF_SRC=$(CHIMESFF_LOC)/chimesFF.cpp
+CHIMESFF_HDR=$(CHIMESFF_LOC)/chimesFF.h
 
-chimescalc.o : $(CHIMESFF_SRC)
+chimesFF.o : $(CHIMESFF_SRC)
 	$(CXX) -c $(CHIMESFF_SRC) -I $(CHIMESFF_LOC)
 
 SERIAL_LOC=$(C_LOC)/../../src
-SERIAL_SRC=$(SERIAL_LOC)/chimescalc_serial.cpp
-SERIAL_HDR=$(SERIAL_LOC)/chimescalc_serial.h
+SERIAL_SRC=$(SERIAL_LOC)/serial_chimes_interface.cpp
+SERIAL_HDR=$(SERIAL_LOC)/serial_chimes_interface.h
 
-chimescalc_serial.o : $(SERIAL_SRC)
+serial_chimes_interface.o : $(SERIAL_SRC)
 	$(CXX) -c $(SERIAL_SRC) -I $(SERIAL_LOC) -I $(CHIMESFF_LOC)
 
 WRAPPER_LOC=$(C_LOC)/../../api
@@ -25,7 +25,7 @@ chimescalc_serial_C.o : $(WRAPPER_SRC)
 
 
 chimescalc_serial_C.so:
-	$(CXX) -shared -o lib-C_wrapper-serial_interface.so chimescalc_serial_C.o chimescalc_serial.o chimescalc.o
+	$(CXX) -shared -o lib-C_wrapper-serial_interface.so chimescalc_serial_C.o serial_chimes_interface.o chimesFF.o
 
 clean:
 	rm -f *.o
@@ -36,8 +36,8 @@ clean-all:
 	rm -f *.so
 
 all:
-	make chimescalc.o
-	make chimescalc_serial.o
+	make chimesFF.o
+	make serial_chimes_interface.o
 	make chimescalc_serial_C.o
 	make chimescalc_serial_C.so
 	make clean
