+++ serial_interface/tests/run_tests.sh
@@ -43,7 +43,8 @@ API[4]="fortran08"; EXE[4]="fortran08_wrapper-serial_interface" ; XTRA[4]="" #"0
 echo "Running $STYLE tests"
 date
 
-for compile in CMAKE MAKEFILE
+#for compile in CMAKE MAKEFILE
+for compile in MAKEFILE
 do
 	echo "Testing compilation type: $compile"
